+++ test/integration/tests/certify.sh
@@ -55,6 +55,6 @@ tpm2_create -Q -g sha256 -G rsa -u certify.pub -r certify.priv  -C primary.ctx
 
 tpm2_load -Q -C primary.ctx -u certify.pub -r certify.priv -n certify.name -o certify.ctx
 
-tpm2_certify -Q -c primary.ctx -C certify.ctx -g sha256 -a attest.out -s sig.out
+tpm2_certify -Q -C primary.ctx -c certify.ctx -g sha256 -a attest.out -s sig.out
 
 exit 0