+++ CHANGELOG.md
@@ -3,6 +3,10 @@
   * tpm2_nvwrite and tpm2_nvread: when -P is "index" -a is optional and defaults to
     the NV_INDEX value passed to -x.
   * Load TCTI's by SONAME, not raw .so file
+  * tpm2_activatecredential: -e becomes -E
+  * tpm2_activatecredential: -e becomes -E
+  * tpm2_certify: -c and -C are swapped, -k becomes -K
+  * tpm2_createprimary: -K becomes -k
   * tpm2_encryptdecrypt: supports input and output to stdin and stdout respectively.
   * tpm2_create: -g/-G become optional options.
   * tpm2_createprimary: -g/-G become optional options.